fix: Restore `responsive` e2e driver option (#25932)
## **Description**

In https://github.com/MetaMask/metamask-extension/pull/18347, we changed
the driver option `responsive` to `openDevToolsForTabs`. This was
because we had a new test suite that intended to test browser behaviour
related to Service Worker restart on manifest v3 builds. Later changes
to chrome's mv3 implementation meant that service worker no longer
restarted, and those e2e tests no longer exist
(`test/e2e/mv3/service-worker-restart.spec.js`).

However, this change had an unwanted side effect on another test suite
(`test/e2e/tests/metamask-responsive-ui.spec.js`). As @Gudahtt puts it
in a slack message:

> It looks like in this PR, the responsive option in the Chrome
webdriver was replaced with openDevToolsForTabs. This is functionally
what that option did, but this doesn't match the semantic purpose for
why the browser opened the dev tools.

> Browser driver options are shared between the Firefox and Chrome
drivers, so now they don't match. The test suite meant to test our UI in
a small viewport is now passing in the openDevToolsForTabs option, but
this doesn't work for the Firefox browser because it doesn't recognize
that option. So the tests are "passing" but they aren't actually running
the steps that are meant to be under test.

This PR reverts the option to its original name.

---------

Co-authored-by: Mark Stacey <markjstacey@gmail.com>