fix: Fix Trezor signing and connecting (#26882)
## **Description**

Trezor signing and connecting was broken in a recent change that
included an update to the `@trezor/connect-web` library (in #26143). The
patch had to be rewritten as part of that update, but in this rewrite an
underscore was missed in referencing a variable (the patch used
`this.settings` rather than `this._settings`).

Additionally, two more changes have been applied to ensure the modified
settings are used everywhere. One was present in the original patch but
missing from the patch in #26143. The other was not present in the
original patch (#23763) because the affected line was added in the
library update, but it seems equally necessary.

[![Open in GitHub
Codespaces](https://github.com/codespaces/badge.svg)](https://codespaces.new/MetaMask/metamask-extension/pull/26882?quickstart=1)

## **Related issues**

Fixes #26875

## **Manual testing steps**

Follow reproduction steps in #26875

## **Screenshots/Recordings**

### Before

See
https://www.loom.com/share/3e1f3d2cc57247a5b02e0ff8d61babb8?sid=1b17e423-16cf-4260-a1ed-674441e25b3a

### After:


https://github.com/user-attachments/assets/0235077f-99da-4de6-87d0-9fcd7179017c

(I stopped partway through connecting, I was already past the point of
it crashing by then).

## **Pre-merge author checklist**

- [x] I've followed [MetaMask Contributor
Docs](https://github.com/MetaMask/contributor-docs) and [MetaMask
Extension Coding
Standards](https://github.com/MetaMask/metamask-extension/blob/develop/.github/guidelines/CODING_GUIDELINES.md).
- [x] I've completed the PR template to the best of my ability
- [x] I’ve included tests if applicable
- [x] I’ve documented my code using [JSDoc](https://jsdoc.app/) format
if applicable
- [x] I’ve applied the right labels on the PR (see [labeling
guidelines](https://github.com/MetaMask/metamask-extension/blob/develop/.github/guidelines/LABELING_GUIDELINES.md)).
Not required for external contributors.

## **Pre-merge reviewer checklist**

- [ ] I've manually tested the PR (e.g. pull and build branch, run the
app, test code being changed).
- [ ] I confirm that this PR addresses all acceptance criteria described
in the ticket it closes and includes the necessary testing evidence such
as recordings and or screenshots.